Split matching code into several class files. Add new classes for matching based on google sample code
